Fix suggested `round` replacement for negatives.
The proposed function for mimicking Python 2 `round` behavior is incorrect for negative numbers. Example:

```
>>> my_round(0.1)
0.0  # good!

>>> my_round(-0.1)
-1.0  # bad!

>>> round(-0.1)
0  # expected
```

I believe the proposed function tried to support negatives (it used `math.copysign()` to add a negative 0.5), however the use of `math.floor()` is wrong here because it floors toward negative infinity rather than toward zero. The cleanest solution I can come up with is to simply use an if-else.

This feels worthy of adding to the errata, but I wasn't sure what section it should go in so I didn't attempt to add it.